nyaymarg-backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
bleach==6.1.0

# ML & Data
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
nltk==3.8.1
joblib==1.4.2

# Charts & Export
matplotlib==3.9.0
seaborn==0.13.2
reportlab==4.2.0
pillow==10.3.0

# Background tasks
celery==5.4.0
redis==5.0.4

# Config & Utils
pydantic-settings==2.2.1
structlog==24.1.0
python-dotenv==1.0.1

# External API clients
tenacity==8.3.0
cachetools==5.3.3

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
aiosqlite==0.20.0
